validate import_limit
git-svn-id: https://plugins.svn.wordpress.org/typo3-importer/trunk@473197 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ typo3-importer.php
@@ -170,7 +170,7 @@ EOD;
 						{$this->news_custom_order}
 					";
 
-					$limit			= (int) get_t3i_options( 'import_limit' );
+					$limit			= get_t3i_options( 'import_limit' );
 					if ( $limit )
 						$query		.= ' LIMIT ' . $limit;
